test: consolidate link visibility test fixtures
